chore: remove ton addons and refactor ton-v2 sdk
