--- packages/routing-sdk/addons/ton-v2/package.json
@@ -1,51 +1,0 @@
-{
-  "name": "@pancakeswap/routing-sdk-addon-ton-v2",
-  "version": "0.0.1",
-  "description": "Routing sdk addon for pcs Ton",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/pancakeswap/pancake-frontend.git",
-    "directory": "packages/routing-sdk/addons/ton"
-  },
-  "main": "dist/index.js",
-  "module": "dist/index.mjs",
-  "types": "dist/index.d.ts",
-  "private": true,
-  "sideEffects": false,
-  "scripts": {
-    "build": "tsup",
-    "dev": "tsup --watch",
-    "test": "vitest --run",
-    "test:u": "vitest --run -u",
-    "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist",
-    "typechecks": "tsc --noEmit -p ./tsconfig.json"
-  },
-  "files": [
-    "dist"
-  ],
-  "peerDependencies": {
-    "@pancakeswap/routing-sdk": "workspace:*"
-  },
-  "dependencies": {
-    "@pancakeswap/chains": "workspace:*",
-    "@pancakeswap/swap-sdk-core": "workspace:*",
-    "@pancakeswap/ton-v2-sdk": "workspace:*",
-    "@ton/core": "^0.59.0",
-    "lodash": "^4.17.21",
-    "tiny-invariant": "^1.3.0"
-  },
-  "devDependencies": {
-    "@pancakeswap/routing-sdk": "workspace:*",
-    "@pancakeswap/tsconfig": "workspace:*",
-    "@types/lodash": "^4.14.168",
-    "tsup": "^6.7.0"
-  },
-  "exports": {
-    "./package.json": "./package.json",
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.mjs",
-      "require": "./dist/index.js"
-    }
-  }
-}
