--- packages/routing-sdk/addons/ton-v2/src/constants/gasCost.ts
@@ -1,7 +1,0 @@
-export const COST_PER_EXTRA_HOP_STABLE_SWAP = 70000n
-
-// Constant cost for doing any swap regardless of pools.
-export const BASE_SWAP_COST_TON_V2 = 50000n // 115000, bumped up by 20_000
-
-// Constant per extra hop in the route.
-export const COST_PER_EXTRA_HOP_TON_V2 = 50000n // 20000, bumped up by 30_000
