--- packages/routing-sdk/addons/ton-v2/src/constants/index.ts
@@ -1,1 +1,0 @@
-export * from './poolType'
