--- packages/routing-sdk/addons/ton-v2/src/constants/poolType.ts
@@ -1,1 +1,0 @@
-export const TON_POOL_TYPE = 'TON_V2' as const
