--- packages/routing-sdk/addons/ton-v2/src/index.test.ts
@@ -1,11 +1,0 @@
-import { expect, test } from 'vitest'
-import * as exportedNameSpaces from './index'
-
-test('exports', () => {
-  expect(Object.keys(exportedNameSpaces)).toMatchInlineSnapshot(`
-    [
-      "TON_POOL_TYPE",
-      "createTonPool",
-    ]
-  `)
-})
