--- packages/routing-sdk/addons/ton-v2/src/index.ts
@@ -1,2 +1,0 @@
-export { TON_POOL_TYPE } from './constants'
-export { createTonPool } from './createTonPool'
