--- packages/routing-sdk/addons/ton-v2/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "@pancakeswap/tsconfig/base",
-  "include": ["./src"],
-  "exclude": ["**/*.test.ts", "./dist/**"],
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./dist",
-    "target": "es2020",
-    "esModuleInterop": true,
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "skipLibCheck": true
-  }
-}
