+++ packages/ton-v2-sdk/src/constants/index.ts
@@ -1,9 +1,4 @@
-export * from './currency/currency'
-export * from './currency/Token'
-export * from './currency/Native'
 export * from './opcodes'
-export * from './CurrencyAmount'
-export * from './Price'
 export * from './contracts'
 export * from './ton.enums'
 export * from './ton.types'
