+++ packages/ton-v2-sdk/src/currency/AgnosticNativeCurrency.ts
