+++ packages/ton-v2-sdk/src/currency/currency.ts
