+++ packages/ton-v2-sdk/src/fractions/CurrencyAmount.ts
@@ -3,8 +3,7 @@ import invariant from 'tiny-invariant'
 import _Big from 'big.js'
 // @ts-ignore
 import toFormat from 'toformat'
-import { Currency } from './currency/currency'
-import { Token } from './currency/Token'
+import { Currency, Token } from '../currency'
 
 const Big = toFormat(_Big)
