+++ packages/ton-v2-sdk/src/utils/getAddress.ts
@@ -1,5 +1,5 @@
 import { Address } from '@ton/core'
-import { Currency } from '../constants'
+import { Currency } from '../currency'
 
 // todo:@eric wait for ton's currency type implemented
 export const getPairAddress = (tokenA: Currency, tokenB: Currency): Address => {
